authorization: remove resourceType from AuthorizeRequest
Since we have resource type on individual permissions now, it doesn't
seem to make sense to have it on the request too - it's just an extra
thing to validate.

As part of this change, I switched to validating that conditions in the
response from the permission handler have the same resource type as the
permission supplied to the handler does - this adds some protection
against implementation errors in handlers (which we don't control).

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>